+++ packages/itwinui-react/src/core/Table/cells/EditableCell.tsx
@@ -2,9 +2,9 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-import React from 'react';
-import { CellRendererProps } from 'react-table';
-import { getRandomValue } from '../../utils';
+import * as React from 'react';
+import type { CellRendererProps } from 'react-table';
+import { getRandomValue } from '../../utils/index.js';
 
 export type EditableCellProps<T extends Record<string, unknown>> =
   CellRendererProps<T> & {