+++ packages/itwinui-react/src/core/Textarea/index.ts
@@ -2,6 +2,6 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-export { Textarea } from './Textarea';
-export type { TextareaProps } from './Textarea';
+export { Textarea } from './Textarea.js';
+export type { TextareaProps } from './Textarea.js';
 export default './Textarea';