+++ packages/itwinui-react/src/core/Tooltip/Tooltip.tsx
@@ -2,9 +2,10 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-import React from 'react';
+import * as React from 'react';
 import cx from 'classnames';
-import { useTheme, CommonProps, Popover, PopoverProps } from '../utils';
+import { useTheme, Popover } from '../utils/index.js';
+import type { CommonProps, PopoverProps } from '../utils/index.js';
 import '@itwin/itwinui-css/css/tooltip.css';
 
 export type TooltipProps = {