+++ packages/itwinui-react/src/core/Tree/TreeNodeExpander.tsx
@@ -2,10 +2,11 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-import React from 'react';
+import * as React from 'react';
 import cx from 'classnames';
-import { SvgChevronRight } from '../utils';
-import { IconButton, IconButtonProps } from '../Buttons/IconButton';
+import { SvgChevronRight } from '../utils/index.js';
+import { IconButton } from '../Buttons/IconButton/index.js';
+import type { IconButtonProps } from '../Buttons/IconButton/index.js';
 import '@itwin/itwinui-css/css/tree.css';
 
 export type TreeNodeExpanderProps = {