+++ packages/itwinui-react/src/core/utils/components/index.ts
@@ -2,15 +2,16 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-export * from './Popover';
-export * from './Resizer';
-export * from './FocusTrap';
-export * from './InputContainer';
-export * from './WithCSSTransition';
-export * from './MiddleTextTruncation';
-export * from './VirtualScroll';
-export * from './VisuallyHidden';
-export * from './Icon';
-export * from './Flex';
-export * from './Divider';
-export * from './LinkAction';
+export * from './Popover.js';
+export * from './Resizer.js';
+export * from './FocusTrap.js';
+export * from './InputContainer.js';
+export * from './WithCSSTransition.js';
+export * from './MiddleTextTruncation.js';
+export * from './VirtualScroll.js';
+export * from './VisuallyHidden.js';
+export * from './Icon.js';
+export * from './Flex.js';
+export * from './Divider.js';
+export * from './LinkAction.js';
+export * from './AutoclearingHiddenLiveRegion.js';